backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
docopt==0.6.2
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-Testing==0.8.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-flask==1.3.0
pytest-watch==4.2.0
python-dotenv==1.1.1
watchdog==6.0.0
Werkzeug==3.1.3
